fix conflict candidate

+++ packages/storage/sqldb/blockchain.go
@@ -20,7 +20,7 @@ type BlockChain struct {
 	NodePosition   int64  `gorm:"not null"`
 	Time           int64  `gorm:"not null"`
 	Tx             int32  `gorm:"not null"`
-	ConsensusMode  int8   `gorm:"not null"`
+	ConsensusMode  int32  `gorm:"not null"`
 	CandidateNodes []byte `gorm:"not null"`
 }
