+++ packages/storage/sqldb/info_block.go
@@ -20,7 +20,7 @@ type InfoBlock struct {
 	CurrentVersion string `gorm:"not null"`
 	Sent           int8   `gorm:"not null"`
 	RollbacksHash  []byte `gorm:"not null"`
-	ConsensusMode  int8   `gorm:"not null"`
+	ConsensusMode  int32  `gorm:"not null"`
 	CandidateNodes []byte `gorm:"not null"`
 }
